requirements.txt
djangorestframework
mongoengine==0.18.2
python-dateutil
Pyjwt
Django==2.2.4
django-cors-headers
nose
pymongo==3.9.0
six==1.10.0
flake8
flake8-import-order
Sphinx==1.5.5
sphinx-rtd-theme==0.2.4
tensorflow==1.4.0
keras==2.1.0
numpy==1.18.1
pillow==7.0.0
scipy==1.4.1
h5py==2.10.0
matplotlib==3.1.2
opencv-python==4.1.2.30
imageai==2.1.0